+++ mmg/exceptions.py
@@ -1,7 +1,4 @@
+class BadConfigError(Exception):
+    """Raised when the configuration is invalid."""
 
-
-class MmgException(Exception):
-    """Base exception for all exceptions raised by mmg."""
-
-    def __init__(self, *args: object) -> None:
-        super().__init__(*args)
+    pass